Update Humidity_monitor.ino
Sensor ranges are included in the Sensor Range structure. This makes it easy to control all three sensors simultaneously and avoid duplicating variables.

The source code is divided into separate functions for each part of the functionality: one function outputs the raw values from the sensors, the second converts them into a percentage of humidity and outputs it as well.

A loop has been added for configuring inputs, which reduces code duplication and makes it easier to support a large number of sensors.

Fewer variables and more clarity in the code structure can reduce RAM consumption and make it easier to understand program logic.